+++ web-admin/src/main/java/org/openhubframework/openhub/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Module of core OpenHub configuration/settings. 
+ */
+package org.openhubframework.openhub.config;
